chore(contracts): compile using cmake
